[refactor]: api 폴더구조 변경
